api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
anthropic==0.42.0
httpx==0.28.0
web3==7.6.0
python-dotenv==1.0.1
pydantic==2.10.0
